requirements/ci.txt
-r base.txt
coverage==3.7.1
django-jenkins==0.16.4